separated context menu control from main menu control
